refactor: move PageEdit to own tsx file
